+++ CHANGELOG.md
@@ -7,4 +7,4 @@ From v2.0.0 to v2.0.1
 ## Changes
 
 ## New Features
-
+- raise error if search space for optuna optimization is missing search space type annotation or if type not in list